refactor(services): migrate base_resolution_usecase to pr/base_resolution
- Move base_resolution_usecase.py → pr/base_resolution.py
- Update imports in pr/create.py and services/__init__.py

Part of issue #2586: Phase 3-cleanup
