+++ src/vibe3/services/pr/base_resolution.py
