+++ src/vibe3/services/pr/create.py
@@ -10,8 +10,8 @@ from rich.prompt import Prompt
 from vibe3.clients import AIClient, AISuggestionClient, BaseResolver
 from vibe3.config import VibeConfig
 from vibe3.prompts import resolve_prompts_path
-from vibe3.services.base_resolution_usecase import BaseResolutionUsecase
 from vibe3.services.flow.service import FlowService
+from vibe3.services.pr.base_resolution import BaseResolutionUsecase
 from vibe3.services.task.binding_guard import ensure_task_issue_bound
